Added difficulty progression
The code is messy (again)
But I added some features that adds some difficulty progression to the
game!

-> Added levels. Obviously, it increments for each level.
-> Changed the way the ghosts react when they are eaten by a powered-up
pacman, instead of resetting all their values, I just teleport them just
outside the ghost pen.
-> Removed debug texts
-> Changed the UI a bit, I just moved the texts to make them look
prettier
-> I think I also fixed the rare bug that happens when pacman eats a
ghost, sometimes when pacman does it, the ghost "freezes" on their
position and it lasts until the entire round. I think it had something
to do with fully resetting their values instead of just teleporting them
outside the ghost pen.

This should be v1.1

+++ src/game_objects.rs
@@ -16,7 +16,7 @@ pub struct PacMan {
     pub speed: f32,
     pub powered_up: bool,
     pub colliding: bool,
-    pub power_up_timer: u32,
+    pub power_up_timer: u16,
     pub score: u32,
     pub lives: u8,
 }
